+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Version: 1.3.1
 - Rubber trees spawn chance config now works
 - Singleblock machines can no longer run a recipe beyond its tier
 - Fluid Hatch locking properly saves between world loads
+- Color-changing tooltip now changes colors properly
 - Fixed recipe conflicts in ore recipes
 - Fixed Drill GUI not updating after 1st click
 - Fixed batteries always working as ULV once charged